requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
praw==7.7.0
prawcore==2.3.0
python-dotenv==1.0.0
requests==2.28.2
schedule==1.1.0
update-checker==0.18.0
urllib3==1.26.15
websocket-client==1.5.1
